测试前端联动

+++ apps/comfy_app/src/node_type.rs
@@ -123,7 +123,10 @@ impl NodeType {
 
     /// 添加输入槽位
     ///
-    /// TODO 参数类型待定
+    /// Args:
+    /// * name: 输入槽位名称
+    /// * r#type: 输入槽位类型
+    /// * extra_info: 额外信息
     pub fn add_input(&self, name: &str, r#type: &str, extra_info: JsValue) -> Result<(), JsValue> {
         let add_input_fn = Reflect::get(&self.inner, &"addInput".into())?.dyn_into::<Function>()?;
 
@@ -163,46 +166,46 @@ impl NodeType {
     // }
 
     /// 获取输入槽位
-    pub fn get_input(&self, index: usize) -> Result<Input, JsValue> {
+    pub fn get_input(&self, index: u32) -> Result<Input, JsValue> {
         let inputs = self.inputs()?;
-        let input = inputs.get(index as u32);
+        let input = inputs.get(index);
         Input::from_js(input)
     }
 
     /// 获取输出槽位
-    pub fn get_output(&self, index: usize) -> Result<Output, JsValue> {
+    pub fn get_output(&self, index: u32) -> Result<Output, JsValue> {
         let outputs = self.outputs()?;
-        let output = outputs.get(index as u32);
+        let output = outputs.get(index);
         Output::from_js(output)
     }
 
     /// 获取小部件
-    pub fn get_widget(&self, index: usize) -> Result<Widget, JsValue> {
+    pub fn get_widget(&self, index: u32) -> Result<Widget, JsValue> {
         let widgets = self.widgets()?;
-        let widget = widgets.get(index as u32);
+        let widget = widgets.get(index);
         Widget::from_js(widget)
     }
 
     /// 设置输入槽位
-    pub fn set_input(&mut self, index: usize, input: Input) -> Result<(), JsValue> {
+    pub fn set_input(&mut self, index: u32, input: Input) -> Result<(), JsValue> {
         let inputs = self.inputs()?;
-        inputs.set(index as u32, serde_wasm_bindgen::to_value(&input)?);
+        inputs.set(index, serde_wasm_bindgen::to_value(&input)?);
         self.set_widgets(&inputs)?;
         Ok(())
     }
 
     /// 设置输出槽位
-    pub fn set_output(&mut self, index: usize, output: Input) -> Result<(), JsValue> {
+    pub fn set_output(&mut self, index: u32, output: Input) -> Result<(), JsValue> {
         let outputs = self.outputs()?;
-        outputs.set(index as u32, serde_wasm_bindgen::to_value(&output)?);
+        outputs.set(index, serde_wasm_bindgen::to_value(&output)?);
         self.set_widgets(&outputs)?;
         Ok(())
     }
 
     /// 设置小部件
-    pub fn set_widget(&mut self, index: usize, widget: Widget) -> Result<(), JsValue> {
+    pub fn set_widget(&mut self, index: u32, widget: Widget) -> Result<(), JsValue> {
         let widgets = self.widgets()?;
-        widgets.set(index as u32, serde_wasm_bindgen::to_value(&widget)?);
+        widgets.set(index, serde_wasm_bindgen::to_value(&widget)?);
         self.set_widgets(&widgets)?;
         Ok(())
     }
@@ -211,9 +214,9 @@ impl NodeType {
 /// 实验性的接口
 impl NodeType {
     /// 设置小部件值
-    pub fn set_widget_value(&self, index: usize, value: f64) -> Result<(), JsValue> {
+    pub fn set_widget_value(&self, index: u32, value: f64) -> Result<(), JsValue> {
         let widgets = self.widgets()?;
-        let widget = widgets.get(index as u32);
+        let widget = widgets.get(index);
 
         if !widget.is_undefined() {
             Reflect::set(&widget, &"value".into(), &JsValue::from_f64(value))?;
@@ -223,9 +226,9 @@ impl NodeType {
     }
 
     /// 设置小部件选项最大值
-    pub fn set_widget_option_max(&self, index: usize, max: f64) -> Result<(), JsValue> {
+    pub fn set_widget_option_max(&self, index: u32, max: f64) -> Result<(), JsValue> {
         let widgets = self.widgets()?;
-        let widget = widgets.get(index as u32);
+        let widget = widgets.get(index);
 
         if !widget.is_undefined() {
             let options = Reflect::get(&widget, &"options".into())?;
@@ -236,9 +239,9 @@ impl NodeType {
     }
 
     /// 设置输出槽位类型
-    pub fn set_output_type(&self, index: usize, slot_type: &str) -> Result<(), JsValue> {
+    pub fn set_output_type(&self, index: u32, slot_type: &str) -> Result<(), JsValue> {
         let outputs = self.outputs()?;
-        let output = outputs.get(index as u32);
+        let output = outputs.get(index);
 
         if !output.is_undefined() {
             Reflect::set(&output, &"type".into(), &JsValue::from_str(slot_type))?;
@@ -248,9 +251,9 @@ impl NodeType {
     }
 
     /// 设置输出槽位标签
-    pub fn set_output_label(&self, index: usize, label: &str) -> Result<(), JsValue> {
+    pub fn set_output_label(&self, index: u32, label: &str) -> Result<(), JsValue> {
         let outputs = self.outputs()?;
-        let output = outputs.get(index as u32);
+        let output = outputs.get(index);
 
         if !output.is_undefined() {
             Reflect::set(&output, &"label".into(), &JsValue::from_str(label))?;
@@ -260,9 +263,9 @@ impl NodeType {
     }
 
     /// 设置输出槽位名称
-    pub fn set_output_name(&self, index: usize, name: &str) -> Result<(), JsValue> {
+    pub fn set_output_name(&self, index: u32, name: &str) -> Result<(), JsValue> {
         let outputs = self.outputs()?;
-        let output = outputs.get(index as u32);
+        let output = outputs.get(index);
 
         if !output.is_undefined() {
             Reflect::set(&output, &"name".into(), &JsValue::from_str(name))?;
@@ -272,9 +275,9 @@ impl NodeType {
     }
 
     /// 设置输入槽位类型
-    pub fn set_input_type(&self, index: usize, slot_type: &str) -> Result<(), JsValue> {
+    pub fn set_input_type(&self, index: u32, slot_type: &str) -> Result<(), JsValue> {
         let inputs = self.inputs()?;
-        let input = inputs.get(index as u32);
+        let input = inputs.get(index);
 
         if !input.is_undefined() {
             Reflect::set(&input, &"type".into(), &JsValue::from_str(slot_type))?;
@@ -284,9 +287,9 @@ impl NodeType {
     }
 
     /// 设置输入槽位名称
-    pub fn set_input_name(&self, index: usize, name: &str) -> Result<(), JsValue> {
+    pub fn set_input_name(&self, index: u32, name: &str) -> Result<(), JsValue> {
         let inputs = self.inputs()?;
-        let input = inputs.get(index as u32);
+        let input = inputs.get(index);
 
         if !input.is_undefined() {
             Reflect::set(&input, &"name".into(), &JsValue::from_str(name))?;
